requirements.txt
asgiref==3.6.0
Babel==2.10.3
blinker==1.5
certifi==2022.9.24
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
colorzero==2.0
cryptography==38.0.4
distro==1.8.0
Flask==2.2.2
gpiozero==2.0.1
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
lgpio==0.2.2.0
MarkupSafe==2.1.2
pigpio==1.78
pycryptodomex==3.11.0
pyinotify==0.9.6
pyOpenSSL==23.0.0
python-apt==2.6.0
python-dotenv==0.21.0
pytz==2022.7.1
requests==2.28.1
rpi-lgpio==0.6
simplejson==3.18.3
six==1.16.0
smbus2==0.4.2
spidev==3.5
ssh-import-id==5.10
toml==0.10.2
urllib3==1.26.12
Werkzeug==2.2.2
